+++ pom.xml
@@ -11,7 +11,7 @@
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     <maven.compiler.source>1.8</maven.compiler.source>
     <maven.compiler.target>1.8</maven.compiler.target>
-    <main.class>ch.heig.gen.MainCommand</main.class>
+    <main.class>ch.heig.gen.commands.MainCommand</main.class>
   </properties>
   <dependencies>
     <dependency>
