+++ src/main/java/ch/heig/gen/commands/BuildCommand.java
@@ -1,15 +1,15 @@
-package ch.heig.gen;
+package ch.heig.gen.commands;
 
 import picocli.CommandLine;
 
 @CommandLine.Command(
-        name = "clean",
+        name = "build",
         description = "Not implemented",
         mixinStandardHelpOptions = true,
         version = "1")
-public class CleanCommand implements Runnable {
+public class BuildCommand implements Runnable {
     @Override
     public void run() {
-        System.out.println("'clean' command not implemented");
+        System.out.println("'build' command not implemented");
     }
 }
