+++ src/main/java/ch/heig/gen/commands/MainCommand.java
@@ -1,4 +1,4 @@
-package ch.heig.gen;
+package ch.heig.gen.commands;
 
 import picocli.CommandLine;
