+++ src/main/java/ch/heig/gen/commands/NewCommand.java
@@ -1,15 +1,15 @@
-package ch.heig.gen;
+package ch.heig.gen.commands;
 
 import picocli.CommandLine;
 
 @CommandLine.Command(
-        name = "build",
+        name = "new",
         description = "Not implemented",
         mixinStandardHelpOptions = true,
         version = "1")
-public class BuildCommand implements Runnable {
+public class NewCommand implements Runnable {
     @Override
     public void run() {
-        System.out.println("'build' command not implemented");
+        System.out.println("'new' command not implemented");
     }
 }
